main.py
from dataclasses import dataclass
from typing import List, Dict
import requests
from bs4 import BeautifulSoup
import pandas as pd
from logging_config import setup_logging

# Configure logging
logger = setup_logging()

@dataclass
class Sneaker:
    name: str
    retail_price: float
    resale_price: float
    profit_margin: float
    sales_volume: int
    last_updated: str

class ArbitrageScout:
    def __init__(self):
        self.config = Config()
        self.retailers = {r.name: r.base_url for r in self.config.retailers}
        self.resale_platforms = {p.name: p.base_url for p in self.config.resale_platforms}
        
    def scrape_retail_data(self) -> Dict[str, List[Sneaker]]:
        """Scrape data from all retail websites"""
        retail_data = {}
        for retailer, url in self.retailers.items():
            try:
                logger.info(f"Scraping {retailer}...")
                response = requests.get(url)
                soup = BeautifulSoup(response.content, 'html.parser')
                # Implement retailer-specific scraping logic
                # ...
                retail_data[retailer] = []
            except Exception as e:
                logger.error(f"Error scraping {retailer}: {str(e)}")
        return retail_data

    def scrape_resale_data(self) -> Dict[str, List[Sneaker]]:
        """Scrape data from resale platforms"""
        resale_data = {}
        for platform, url in self.resale_platforms.items():
            try:
                logger.info(f"Scraping {platform}...")
                # Implement platform-specific scraping logic
                # ...
                resale_data[platform] = []
            except Exception as e:
                logger.error(f"Error scraping {platform}: {str(e)}")
        return resale_data

    def calculate_profit_margins(self, retail_data: Dict, resale_data: Dict) -> List[Sneaker]:
        """Calculate profit margins and filter opportunities"""
        opportunities = []
        # Implement profit margin calculation logic
        # ...
        return opportunities

    def generate_reports(self, opportunities: List[Sneaker]):
        """Generate reports and notifications"""
        # Implement reporting logic
        # ...

if __name__ == "__main__":
    scout = ArbitrageScout()
    retail_data = scout.scrape_retail_data()
    resale_data = scout.scrape_resale_data()
    opportunities = scout.calculate_profit_margins(retail_data, resale_data)
    scout.generate_reports(opportunities)
