requirements.txt
# Core Dependencies
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Logging
loguru==0.7.0

# Data Processing
numpy==1.25.2

# Documentation
mkdocs==1.5.2
mkdocs-material==9.2.7
